refactor: move ClientTraceIdSimpleFormatter class to logging package
